move new classes to chat package remove old classes
