--- src/main/java/com/faforever/client/chat/CategoryOrChatUserListItem.java
@@ -1,29 +1,0 @@
-package com.faforever.client.chat;
-
-import lombok.Value;
-
-/**
- * Represents either a chat user or a chat user category.
- * <p>
- * All attempts to use a tree table view failed (bugs everywhere) so it was decided to use a simple list view that
- * contains "category" items.
- * <p>
- * If it's a category-only object, {@code user} will be {@code null}. If it's a chat user, {@code category} will be
- * {@code null}.
- */
-@Value
-class CategoryOrChatUserListItem {
-
-  ChatUserCategory category;
-  ChatChannelUser user;
-
-  CategoryOrChatUserListItem(ChatUserCategory category) {
-    this.category = category;
-    this.user = null;
-  }
-
-  CategoryOrChatUserListItem(ChatChannelUser user, ChatUserCategory category) {
-    this.user = user;
-    this.category = category;
-  }
-}
