+++ src/main/java/com/faforever/client/chat/ChatUserItem.java
@@ -1,8 +1,5 @@
-package com.faforever.client.chat.test;
+package com.faforever.client.chat;
 
-import com.faforever.client.chat.ChatChannelUser;
-import com.faforever.client.chat.ChatUserCategory;
-import com.faforever.client.chat.ChatUserItemController;
 import com.faforever.client.theme.UiService;
 import javafx.scene.Node;
 import lombok.RequiredArgsConstructor;
