+++ src/main/java/com/faforever/client/chat/ListItem.java
@@ -1,7 +1,5 @@
-package com.faforever.client.chat.test;
+package com.faforever.client.chat;
 
-import com.faforever.client.chat.ChatChannelUser;
-import com.faforever.client.chat.ChatUserCategory;
 import com.faforever.client.theme.UiService;
 import javafx.scene.Node;
 import org.fxmisc.flowless.Cell;
